Backups/Backup13-With Agent/requirements.txt
# Core dependencies
playwright>=1.40.0
python-dotenv==1.0.0
openai>=1.30.0
groq==0.4.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Image processing
Pillow==10.1.0
pytesseract==0.3.10

# UI dependencies
dash[async]>=2.17.0
dash-bootstrap-components>=1.5.0

# Stealth mode
playwright-stealth>=1.0.0

# Async support
aiofiles>=23.0.0
